requirements.txt
fastapi>=0.115.0
geoip2>=4.8.0
httpx>=0.27.0
PyYAML>=6.0.2
uvicorn>=0.30.0
